+++ changelog.md
@@ -9,6 +9,9 @@
 * Suppress warnings like:
   `Warning: ../../../gobject/gsignal.c:2685: instance '0x29948460' has no handler with id '117'`
   when processes finish
+* Suppress warnings like:
+  `Warning: value "100" of type 'gint' is invalid or out of range for property 'page-number-start' of type 'gint'`
+  when importing more than 99 images
 
 
 ## 3.0.1 (2026-03-15)